:ok_hand: Use <slot>
