--- src/components/buttons/TYPES.js
@@ -1,10 +1,0 @@
-export const TYPES = {
-  DEFAULT: '',
-  PRIMARY: 'primary',
-  SECONDARY: 'secondary',
-  TERTIARY: 'tertiary',
-  FLAT: 'flat',
-  GHOST: 'ghost',
-};
-
-export const ALL = Object.values(TYPES);
