+++ src/stories/HdButton.stories.js
@@ -1,8 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import { storiesOf } from '@storybook/vue';
 import { text, select } from '@storybook/addon-knobs';
-import { ALL, TYPES } from 'hd-blocks/components/buttons/TYPES';
-import HdButton from 'hd-blocks/components/buttons/HdButton.vue';
+import HdButton, { TYPES } from 'hd-blocks/components/buttons/HdButton.vue';
 
 storiesOf('HdButton', module)
   .add('Playground 🎛', () => ({
@@ -14,13 +13,12 @@ storiesOf('HdButton', module)
       },
       modifier: {
         type: String,
-        default: select('Modifier', ALL, TYPES.default),
+        default: select('Modifier', Object.values(TYPES), TYPES.default),
       },
     },
     template: `
       <HdButton
-        :text=text
         :modifier=modifier
-      />
+      >{{ text }}</HdButton>
     `,
   }));
